% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{playlist}
\alias{playlist}
\title{Create a playlist}
\usage{
playlist(tracks = list())
}
\arguments{
\item{tracks}{A list of track objects}
}
\value{
An S3 object of class "playlist"
}
\description{
Create a playlist
}
