% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Create a music track}
\usage{
track(title, artist, duration, bpm)
}
\arguments{
\item{title}{Track title}

\item{artist}{Artist name}

\item{duration}{Duration in seconds}

\item{bpm}{Beats per minute}
}
\value{
An S3 object of class "track"
}
\description{
Create a music track
}
